This contains the interface for the Board
The board class will contain the logic for holding player co-ordinates. get player Location returns a Position, which will have either: the location string as "Outside" and x & y co-ordinates, or the location string will contain the name of the estate they're in and null for x & y. function implementations to come, I'm trying to do as much as I can in umple, but if y'all need results asap I can build them in java and get a working version out to you asap